Convert the popups pyauto test to browser_tests.
testPopupBlockedEverySec: I didn't port this since it seemed redundant with other tests, and it's permanently disabled anyways.
testPopupBlockerEnabled is a duplicate of a subset of testLaunchBlockedPopup. same with the incognito version
testAllowPopupsFromExternalSite was really just testing that an exception works, not that it's going to an external networkw which we don't want for tests. I rewrote it to use a local url.
testBlockedPopupNotShowInHistory is testing something that's not the current behavior, which is why it's disabled. if this changes, someone can add a test for the new behavior then

BUG=143637,83827,47935
Review URL: https://chromiumcodereview.appspot.com/10913043