Upstream Android IME support (ImeAdapter).
This CL contains most of the ImeAdapter code that allows Android IMEs to work with Chrome. The
key class is ImeAdapter on the Android/Java side, which has an inner class AdapterInputConnection
which implements Android's inputConnection interface. Also important is the method
ContentViewCore#imeUpdateAdapter, which is called in response to ViewHostMsg_TextInputStateChanged
messsages from the renderer.

This CL also adds SelectionHandleController and InsertionHandleController, but only as stubs.

BUG=136691


Review URL: https://chromiumcodereview.appspot.com/10911012