Extension Commands: Make sure we can open the Configure Commands overlay from the install bubble.
When an extension with an Extension Command is installed, we show an installation bubble. We want to add to that bubble a "Configure shortcut" link that takes you directly to the Extensions page with the Configure Command overlay open. We have infrastructure for doing this in the Settings page, but that infrastructure isn't there for the Extensions page. This buys us time, allowing us to complete the installation bubble link.

BUG=145951
TEST=None
Review URL: https://chromiumcodereview.appspot.com/10911032