Make IndexedDB layout browser tests save incorrect results to a file
Like it says in the bug, this won't help poring over a log on a bot, you'd still have to copy/paste and then unescape those results manually. But outputting to a file is so helpful when running a small amount of tests locally in a develop-compile-test cycle.

The expected and actual results of the test that inspired this are each >3k, of which about 100 bytes differed.

BUG=144786

Review URL: https://chromiumcodereview.appspot.com/10869061