Fix a bug in downloads.js about download ids.
In Downloads.prototype.isUpdateNeeded(downloads), |downloads| is an array of objects (associate arrays) sent by DownloadsDOMHandler::SendCurrentDownloads(), whereas |this.downloads_| is an associate array mapping download ids (as in DownloadItem::GetId()) to |Download| objects (similar to but slightly different from the associative arrays sent by SendCurrentDownloads()). This code used to assume that |Download.id_| was in the same counter space as the indices into |downloads|, but that is no longer the case. isUpdateNeeded() should explicitly use |downloads[i].id| as keys into |this.downloads_|.

This really only caused problems when there are >150 downloads and the user deletes one. In this case, DownloadsDOMHandler sends 150 downloads to downloadsList(), not including the removed item, but including a different item. This caused isUpdateNeeded to look for |downloads[150]|, which is impossible because DownloadsDOMHandler truncates |downloads| to 150 items.

BUG=146020

Review URL: https://chromiumcodereview.appspot.com/10909064