Merge 154234 - Prevent access to chrome.app.* APIs from extensions
This is still sort of a band-aid solution to the problem, but should be enough
to solve the ReleaseBlock issue. The deeper solution to getting rid of the 
RemoveDisallowedAPIs function in extension_api.cc should address this case as well. 

BUG=141335


Review URL: https://chromiumcodereview.appspot.com/10855164

TBR=asargent@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10916122