Fix bug in buildbot's mergeRequests handling.
"For what I understand", this was because it was called with 2 arguments instead
of 3, so it was throwing an exception, which was returned by getResult(). Since
the code in process/builder.py was doing a "if wfd.getResult():", this was
passing.

Note that the bug was already fixed upstream, so not adding anything to
buildbot_8_4p1/README.chromium.

Still, using lambda *_: False as a safety measure since it's work fine with
either 2 or 3 arguments.

TBR=szager@chromium.org,rcui@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/10912064