Merge 152730 - Retry channel setup for NPAPI plugins.
For the referenced bugs, assume that an undiscovered Mr X is closing
the channels, but that the closing is not 100% persistent.  So retry
the setup.

For cases where everything works, this is a non-event, the first time
through will get the channel.  It only kicks in for cases where the
plugin is either going to show a failure infobar, or would previously
have crashed (the 97285 case).  So this doesn't seem like it has a
strong downside.

BUG=97285,141055


Review URL: https://chromiumcodereview.appspot.com/10824388

TBR=shess@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10909091