Reland 10907055 which does:
We need to set the session id to the WebRtcAudioDevice to be able to use the correct microphone which is selected by the users via infobar.

This patch also fix the problem that LocalMediaStream.stop does not stop audio recording.

TBR=tommi@chromium.org

BUG=146308, 140444
TEST=manual tests:
using apprtc.appspot.com, make sure the correct microphone is used.
http://neave.com/webcam/html5/

Review URL: https://chromiumcodereview.appspot.com/10910108