Move gpu blacklist to content side.
When it's on the content side, it can be part of the GpuDataManager, so we don't need to make it a Singleton.

Also, this CL did some cleanup related to this moving.

The main reason for this refactoring is because blacklist is tightly coupled with GpuDataManager.  Putting them on the this side makes it much easier to test the code, and the design is much better.

BUG=
TEST=bots
Review URL: https://chromiumcodereview.appspot.com/10908110