Move chrome/browser/ui files to their own target, chrome_browser_ui
This makes chrome_browser.gypi a lot smaller, which helps with
the "ar commandline to long" that some platforms have.

On Mac, rebuild time after touching a file below chrome/ui improves
from 12s to 8s on my MBP while plugged in, and goes from 20s to 14s
while on battery. (This is because only the chrome_browser_ui lib
needs to be rebuilt, instead of all of libbrowser.a)

BUG=none

Review URL: https://chromiumcodereview.appspot.com/10911127