Move webkit/forms/ build into a new webkit_forms.gypi
Most webkit/* directories have their own .gypi file with a static library
building that directories files. This does the same for webkit/forms
instead of the sources being inlined into the glue target.

In addition to be being more like the other directories, this makes it
easy for a build to use a subset of webkit/, which is important for
iOS.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10897002