cros: Add chat icon to app list
Adds an icon for the chat extension to the app list window.
Since this extension is not an app, add a special launch method
that sends an rtcPrivate.onLaunch to any listening extension.
Will require an updated chat app to actually open a window.

BUG=141888
TEST=manual, go to Tools > Extensions, go into developer mode, install chrome/test/data/extensions/api_test/rtc_private/events unpacked extension, click on Chat icon in app list, observe chrome log for "HANDLER: got default action", source: chrome-extension://jdaiaafaoeaejklkkbndacnggmgmpkpa/background.js


Review URL: https://chromiumcodereview.appspot.com/10915135