Make Preferences.registeredPreferences_ a non-static property
Preferences.registeredPreferences_ is defined in the prototype, making
it appear like a regular, non-static property that each instance sets to
its own value. In reality, the property is of a non-POD type and all
instances of the Preferences class end up manipulating the same shared
object.

This CL moves the definition of registeredPreferences_ to the Preferences
constructor, giving each instance an independent object.

Since Preferences is a singleton anyway, the CL is a clean-up without any
functional changes.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10908146