Partial revert of 155693.
In change 155693 we pulled in a new pnacl toolchain that is 32-bit
(where the previous one was 64-bit).

The chromeos chroot does not have the 32-bit version of zlib.
This causes pnacl-ar to blow up.

Other changes to build gio_lib as a .so cause gio_lib.so to elide gio.a.
Preventing an incremental build after a full revert.
Doing a partial revert instead.

R=noelallen@chromium.org
BUG=None
TEST=None

Review URL: https://chromiumcodereview.appspot.com/10911184