events: Inlcude the event-phase and event-result in Event.
An event-handler can be installed on a target as both a pre-target and a post-target handler.
In such cases, it will be necessary to know the phase of the event-dispatch. So add EventPhase.
Also, from a handler, it may be necessary to know whether the event has been handled by any other
handlers before this. So include EventResult in the event too.

BUG=none.

Review URL: https://chromiumcodereview.appspot.com/10916210