Don't consider a Bluetooth adapter present until it has an address.
Since we don't have an observer method that an address has changed,
and one doesn't really make sense, change the previous behavior on
change of a default adapter to inform observers that an adapter was
removed and a new one added rather than pretend nothing happened.

This also makes it consistent with the (more likely) case where the
new adapter address is not yet known, and the observer has to be
informed of the old adapter removal anyway.

BUG=chromium-os:34283
TEST=unit_tests

Review URL: https://chromiumcodereview.appspot.com/10905190