Fix non-multi-profile-aware uses of kBrowserThemeDidChangeNotification.
The kBrowserThemeDidChangeNotification is global and may not apply to
a given view when multi-profiles are used. Fix users of this notification
that blindly use the the new theme to instead use the correct theme for
their windows.

Also, a couple of changes to bookmark_bar_controller.mm to set the
delegate of a button to nil when that button is not added to the
bookmark bar. This matches what's done to buttons that get removed
from the bookmark bar and is good hygiene against zombies.

BUG=147638
TEST=Have two profile windows open. Have the BMB open in profile A. In
profile B, change themes to a dark theme. This should not affect the
window of profile A.

Review URL: https://chromiumcodereview.appspot.com/10915181