Switch the NPAPI test plugin to the Cocoa event model
The Carbon event model is no longer supported, but a special exception
was made for the test plugin. This switches the test plugin to the Cocoa
model and removes the exception. Two changes other than the negotiation
are necessary:
- The test plugin code that interacts with events has been updated to use
  NPCocoaEvent.
- The checks in SetWindow that verify that 'window' is non-NULL have been
  conditionalized; 'window' is always NULL under the Cocoa event model.

All support for the Carbon event model is now vestigial; it will be
removed in a follow-up CL.

BUG=134359


Review URL: https://chromiumcodereview.appspot.com/10915162