Revert 155868 - Merge 155858 - Prevent a 0-time delay in the hung plugin checker.
Fix a bug in the Pepper hung plugin checker where it would post a task to check for hangs with a negative timeout, causing an assertion.

This would result in the immediate running of the task and would cause a loop of checking in the renderer process that would use all the CPU. This condition is hit when we're blocked for longer than the hang threshold, but due to another message coming in from the plugin, haven't yet hit the "hard" hang threshold.

This also adds a fatal assertion for this condition if this patch doesn't fix it, since we want the reports rather than get weird hangs.

BUG=144077

Review URL: https://chromiumcodereview.appspot.com/10914183

TBR=brettw@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10913182

TBR=brettw@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10918173