Improve the tests for GpuDataManagerImpl.
Before we test with calling a bunch of private APIs.  Now we test with calling the public APIs, and the tests are as close to production code path as we can get.

Also, move some logic to the production only code path, so it won't interfere with testing.

BUG=
TEST=content_unittests
Review URL: https://chromiumcodereview.appspot.com/10911192