Add <(SHARED_INTERMEDIATE_DIR)/webkit to webkit_compositor_support target to work in WebKit checkout
In a webkit checkout using the chromium port, chromium files that use includes of the form "third_party/WebKit/.."
work thanks to a gyp step that maps WebKit headers into the shared intermediate dir. For this to work, all
targets that are compiled on the WebKit side in this configuration have to have this path in the include dir.

BUG=

Review URL: https://chromiumcodereview.appspot.com/10909176