Initial NPAPI plugin support in Win Aura.
Scrolling jank is noticable because the plugin moves at a different time than the content is updated. There are also a couple of other functions in RenderWidgetHostViewWin that enumerate plugin Windows that need to be shared, but I'm leaving these for a future change.

The code in render_widget_host_view_base.cc is all moved from render_widget_host_view_win.cc with the exception of the lines between the USE_AURA ifdef.
Review URL: https://codereview.chromium.org/10905122