Allow running browser tests each in single-process mode. This is useful for Android team, which want to run every test in content_browsertests in single-process mode. Browser tests need to run in separate processes though, each since browser initialization can't happen twice.
Currently, if --single-process is specified, the second test run will assert because of the multiple browser-process initialization problem. This change makes it so that it works as expected. However, it will still fail as before if --single-process and --gtest_filter=<multiple_tests> is passed.
Review URL: https://codereview.chromium.org/10928114