Omnibox: Fix HistoryQuick provider can_inline bug
The HistoryQuick provider erroneously sets can_inline to false when the user's input matches mid-word in the scheme, even if the scheme is droppable.  For instance, for the omnibox input "t", the URL http://teams.foo.com/ would be considered not inlineable.  This is wrong.

This changelist revises the can_inline code to make it correct in these cases.

I added tests.  Before this change, the new tests fail.  Specifically, lines 158 and 174 fail.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10907176