Prime the Windows theme DLL earlier.
On Windows 8, we need to check if we are touch-enabled before we load our resources, and this happens before the code to prime the theme DLL gets executed. Since checking for touch-enable requires a call into the theme dll, it didn't get properly initialized in the render process... So moving this code to be done earlier was required...

At the same time, I changed the code a bit to rely on the current window station to no be WinSta0 to decide whether we need to do this or not, as opposed to looking at the sandbox state.

BUG=139241
TEST=Make sure we always get the proper theme for UI controls like scrollbars.


Review URL: https://chromiumcodereview.appspot.com/10914119