Refactor the P2PSocketDispatcher to be created on the RenderThread instead of the RenderView. Removes the use of routing_id.
The reason for this refactoring is that we want to be able to create a PeerConnection on the RenderThread instead of the RenderView.
PeerConnection is dependent on the P2PSocketDispatcher.
PeerConnections are created in renderer_webkitplatformsupport_impl and should not be affilitated with a render view.
 https://chromiumcodereview.appspot.com/10703095/diff/13017/content/renderer/renderer_webkitplatformsupport_impl.cc

BUG=

Review URL: https://chromiumcodereview.appspot.com/10917167