Build the user_agent target for iOS
This enables building just the user_agent target of webkit_support.gyp
on iOS, and adds iOS versions of the UA-construction utility methods.

Eventually the goal is for iOS not to build any code in webkit/, but
user agent is tied in to enough other code in webkit/ that it's
unlikely to be extracted in the short term, so adding minimal iOS
support to webkit_support.gyp is a necessary bridging measure.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10905184