Change the location to notify OnDisplayModeChanged() event.
It was called by PostTask() in SetDisplayMode(), but it's
- run too earlier than the actual timing
- may not run with other methods (such like CycleOutputMode)

Then this CL puts a timer and invokes the notification based on the
RandR events in Dispatch().

This also adds a check to OutputConfiguratorAnimation since
OnDisplayModeChanged() can now happen anytime, not paired to the fade-out.
BUG=145266

Review URL: https://chromiumcodereview.appspot.com/10916249