Fix IOSurfaceReuse test under Mac ASAN bots
The ASAN Mac bots were flakifying the correlation between window resize and IOSurface resize. Dunno why. Changed the test so that it only verifies a create relative to a specific resize call.

Setting TBR because this test seems to fail in other cases with the CQ.

BUG=148633

TBR=zmo

Review URL: https://chromiumcodereview.appspot.com/10908218