Resizes downloaded high resolution wallpaper to small resolution and saves it to file.
This CL resizes the downloaded wallpaper to small resolution and saves it to file.

Note: It will use the high resolution wallpaper regardless the screen size initially. But next
time we load the wallpaper(e.g. logout and login again), it should use the appropriate
resolution wallpaper. The justification is we dont want to block the wallpaper painting
on resize high resolution to small resolution. We want to display the downloaded
wallpaper ASAP.

BUG=147586

Review URL: https://chromiumcodereview.appspot.com/10914177