Move creation of PeerConnection from the RenderView to the RenderThreadImpl.
This is a proposed solution for the problem with using WebFrame* web_frame = WebFrame::frameForCurrentContext() in
renderer_webkitplatformsupport_impl.cc

https://chromiumcodereview.appspot.com/10703095/diff/13017/content/renderer/renderer_webkitplatformsupport_impl.cc

This should be seen as step 1 in refactoring how PeerConnections are created and used. If PeerConnections should be created in the RenderThreadImpl - so should all objects that it is dependent on such as the P2PSocketDispatcher that is currently created and owned by the RenderViewImpl.

BUG=

Review URL: https://chromiumcodereview.appspot.com/10919122