Introduce a couple of abstract bases for WebDataService.
WebDataService is really a collection of different services that have a common initialization and some commonality e.g. how to cancel requests, and it's probably best that the interfaces reflect this logical split so that the implementation might some daybe tidied up.  Starting this off with a WebDataServiceBase for the common bits, and an AutofillWebData for the Autofill-specific functionalities of WebDataService.

Using these interfaces to break Autofill's concrete dependency on WebDataService.

TBR=ben@chromium.org
BUG=140037

Review URL: https://chromiumcodereview.appspot.com/10908065