Fix Instant test flakiness.
See the bug link for details. Though the root cause (GdkEventConfigure
events closing the popup and causing Instant to Hide()) hasn't been
fixed, it seems that calling FocusOmniboxAndWaitForInstant() and
BringBrowserWindowToFront() fixes both the timeouts as well as the
CHECK(event->button.window) failures, as measured empirically. This is
probably due to both those calls spinning the message loop. I've run
this CL on the linux_chromeos and linux_rel bots a hundred times, and on
the win and win_rel bots about 40 times, with just 1 flake over the
hundreds of runs.

Other changes:

+ Notify Instant when the omnibox already has focus; otherwise we wait
  forever for a page load that never happens.

+ Use WindowedNotificationObserver more robustly. Previously, I had
  reasoned that we couldn't miss notifications, since they result from
  receiving renderer messages, which only happens after we return to the
  MessageLoop. But, there are nested MessageLoop::Run() calls hidden all
  over the place, so it's safer to wrap the observation window properly.

+ MAYBE-ify tests so they still compile, though they are disabled.
+ Split DoesNotCommitURLs into separate tests.
+ Check Task Manager prefixes more robustly.

BUG=146419,145754
R=dhollowa@chromium.org
TEST=Instant tests shouldn't flake, on any platform.

Review URL: https://codereview.chromium.org/10905216