Prevent window from being dragged under launcher in dual display mode.
This CL makes a window not to go under launcher while dragging the window in dual display mode.
When displays are positioned in vertical relationship and a window is dragged across laucher,
the window is stick to the work area where  the mouse cursor is located and does not overlap launcher.

BUG=145652


Review URL: https://chromiumcodereview.appspot.com/10933038