[Android] Add a test content provider into android_webview, port content url tests.
As it is required to have a package for a content provider, I decided
to have two separate test content providers for android_webview here and
Chrome for Android downstream.

There is an issue with the provider package. As the package for tests
is called 'org.chromium.android_webview.tests' (not .test), the provider
must be in the same package.

BUG=146009

Review URL: https://chromiumcodereview.appspot.com/10916148