During installation of a packaged app using the serial permission, the install alert should call out the usage of that permission.
Also change serial permission to allow it to be optional.

This one was mildly tricky because it is a dev-only
permission, at least for today. So we had to use
extensions::Feature::ScopedCurrentChannel to get our test
to work.

Pair-programmed with dharcourt, who isn't yet a committer. Adding
him anyway as a reviewer to let him go through the process.

BUG=149018
TEST=added

Review URL: https://chromiumcodereview.appspot.com/10905278