aura: Make sure the root-window retains the X events it was already listening for.
When aura starts listening for resize events on the X root-window, it
needs to make sure it continues to listen to other X events on the
root-window selected earlier.

Also, split out the handling of events targetted for the X root-window
into a separate function.

BUG=148784

Review URL: https://codereview.chromium.org/10913253