Disable presubmit script on webkit/compositor_bindings
Along with cc/, this is the other directory for compositor code that recently
moved from WebKit to Chromium. In the short term, to make it easier for folks
moving patches in flight from WebKit to Chromium (and so that they can use the
commit queue), just ignore these files.  These suppressions will be removed as
soon as the style in both of these directories have been updated.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10918252