Enable Push Messaging for extensions.
Currently the Push Messaging feature only enabled in _permission_features.json as a platform app, we should also be enabled as an extension, since we've promised our partners that we would run as an extension.

TESTING=I tested this with an extension, and it loaded and ran.

BUG=148561


Review URL: https://chromiumcodereview.appspot.com/10915236