Pepper Flash Mac: Fix fullscreen drawing with 3d disabled
If flash 3d is disabled then fullscreen flash would only occupy a quart of the screen.

In the 3D case we scale the view port in RenderWidgetFullscreenPepper::InitContext(). If 3D is disabled we draw using PepperWidget::paint() which doesn't take into account the scale factor.

This CL updates PepperWidget::paint() to scale the canvas if necessary.

BUG=146479
TEST=
Tested the following cases in normal mode and fullscreen:
  - GPU compositing disabled, gpu black list disabled
  - GPU compositing enable, gpu black list disabled
  - GPU compositing disabled, gpu black list enabled
  - GPU compositing enabled, gpu black list enabled
Also tested the above 4 scenarios with a version of the flash plugin with HiDPI support. I verified that everything looks correct on chatroulette.com and youtube.com.
(Note, youtube.com has an issue in HiDPI mode where the player controls are not scaled up, ihf@ is looking into this separately.)

Review URL: https://chromiumcodereview.appspot.com/10919272