Use a single flag for enabling experimental WebKit features.
Merge multiple command line webkit features into a single flag.  This
has the benefit that when a new feature is added, we can add it to this
flag rather than requiring users to toggle another flag.  This is
similar to what extensions does for experimental APIs.

BUG=148503


Review URL: https://chromiumcodereview.appspot.com/10917269