When a DEPS file is modified, run gclient sync.
It runs it by specifying BASE/HEAD to stay at the current revision.
This works around the lack of gclient sync when a DEPS file is applied on the
Try Server when triggered from Rietveld. This was handled my the Try Server
inside scripts/slave/chromium_command.py manually in the case a 'patch' is
specified.

TBR=rogerta@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/10918264