Launch wallpaper manager/picker in a container and hide app icon from app list
According to PM, we want to launch the wallpaper manager in a 640x440 window.
And we also want to fit 5 thumbnails in a row.

BUG=149250
TEST=
launch wallpaper manager from context menu on a chromeos device. Verify it the width and
height is 640x440. (note: the window frame height should also included in the 440)

Review URL: https://codereview.chromium.org/10907235