Forward Capture.
This one is a little trickier, since there are two capture concepts in play:
- aura capture, used by the event dispatch logic in RootWindow to figure out where to send mouse drag events
- win32 capture, used by windows to determine where to send subsequent mouse events.

NativeWidgetPrivate::Set/Get/HasCapture are implemented in NativeWidgetAura setting only the first type, since that's all that's relevant in ash-land, where you don't need to track mouse events outside the HWND that is the RootWindowHost.

Not-so for Desktop Aura, so we need to set native capture too.

Capture can move within the aura::Windows in the RootWindowHost, but regardless we must always keep capture of the HWND associated with the in sync.

BUG=146077
R=scottmg@chromium.org
Review URL: https://codereview.chromium.org/10907258