Merge 157158 - Disable crashing code in the pepper plugin code.
The dispatcher wrapper outlives the render view, so it was an error to save the RenderView pointer and use it later.

This patch will break out-of-process gamepad which is currently unused, as well as the async flash device ID which is also currently unused.

BUG=149850

Review URL: https://codereview.chromium.org/10909273

TBR=brettw@chromium.org
Review URL: https://codereview.chromium.org/10928238