Bluetooth: Defer observer methods for properties until have address
The AdapterPoweredChanged and AdapterDiscoveringChanged observer
methods could be called while the adapter had no address, and was
claimed not to be present, potentially confusing the UI. Defer
until the address is obtained, and the adapter announced as present.

Add lots of property-related unit tests.

BUG=chromium-os:34498
TEST=unit_tests
Review URL: https://codereview.chromium.org/10910309