Browser test for the push messaging feature
We've implemented a push messaging feature and checked it in with a unit test.
This change is adding a browser test to get better coverage in a running
environment.

The browser test basically brings up a small packaged app, and injects a message
into the sync code. For the test to succeed, the message must arrive into the
JS code of the packaged app.

BUG=148550


Review URL: https://chromiumcodereview.appspot.com/10915216