Always include base extension type in Mac select file dialog
Previously, we were only converting the extension to a UTI, and using that to control what files the user could load using the generic select file dialog. For some reason, Mac's conversions of some types - such as '.ogg' - resulted in a UTI prefixed with 'dyn' that didn't seem to correspond back to '.ogg' itself.

The solution here is just to always include the original extension as well as the UTI. FWIW, the rest of this file needs updating to support lots of more complex things we now do with Packaged Apps.

BUG=148840


Review URL: https://chromiumcodereview.appspot.com/10915252