Force-destroy the media pipeline when Chrome OS Audio/Video player closes
This is not a complete fix but a workaround for the bug. The root cause is crbug.com/149957.
This patch resets the media source url to an empty string to make sure that the media pipeline is destroyed even if the document is leaked.

BUG=chrome-os-partner:13484

Review URL: https://codereview.chromium.org/10943008