Android browser-side compositor initialization cleanup.
GraphicsContext becomes a factory for surface handles, so that
we correctly generate one per RWHV. These are the handles used by the
render-side compositors to share the output with the browser-side
compositor.
Then move ownership of the UI output surface handle to the compositor.
The shared context is used for creating the transport textures and
is now owned by the ImageTransportFactory singleton (it is also used
to insert the sync point). We simply use an offscreen context here
since all contexts share resources on Android.
This also avoids races between the UI output surface being ready
and the renderer starting compositing.

TBR=yfriedman@chromium.org for change to OWNERS file

Review URL: https://codereview.chromium.org/10915315