chromeos: Fix eternal cellular activation spinner
Change our state machine to attempt to reconnect after entering OTASP.

While the cellular activation process is far from deterministic, it seems like
we've been hanging on getting out of OTASP because we stop trying to connect to
the provider when we enter OTASP.  At some point, this must have made sense
because the connection manager would automatically try to connect modems to
cellular service under certain conditions.  However, shill does not do this.

This seems to be a regression introduced in:

https://chromiumcodereview.appspot.com/10456045/


BUG=chromeos-partner:10294
TEST=Manually test that activation does not hang

Review URL: https://chromiumcodereview.appspot.com/10913209