Update page icon in the omnibox
The location bar and the omnibox popup currently show an
"empty page" icon for http entries. This CL makes it so that
the asset in the location bar is different from the one in
the popup, and updates the asset used in both places.

BUG=149903

Review URL: https://codereview.chromium.org/10951023