unchecked_malloc() for Skia on OSX.
On tcmalloc-based versions of Chromium, skia resets the global new
handler to prevent the oom-killer from kicking in.  Traditionally the
OSX developers have been agin this, because it's not thread-safe, and
we have questions about whether it's even the right thing to do.  This
change should be thread-safe because g_old_malloc is only set on the
main thread at startup.

As far as whether it's the right thing to do, I'm tired and want to
quit hearing about it.

Additionally, remove a bit of leftover pre-10.6 code.

BUG=103980,73751,117949


Review URL: https://chromiumcodereview.appspot.com/10908245