Add wrapper container for a vector of OwnPtr<T>
In cc we frequently use WTF::Vector<WTF::OwnPtr<T> > to store lists of single-ownership
objects. In WTF this works nicely using template specialization in the container
implementation. In chromium we use STL containers and can't modify the implementation,
so this adds a wrapper to do roughly the same thing.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10940002