Revert 157716 - [Chromoting] Build Chromoting Host installation in two steps: (binaries & .wxs) -> .zip -> .msi.
The two step process is needed so that we could take the produced .zip and use it to sign the binaries and installation on a separate machine. Decompiling .msi using dark proved to be a bad idea due to bugs in dark.exe.

BUG=146863


Review URL: https://chromiumcodereview.appspot.com/10948050

TBR=alexeypa@chromium.org
Review URL: https://codereview.chromium.org/10956010