Convert CC scheduler logic to use base::TimeTicks/Delta instead of doubles
Instead of using double seconds, this converts the scheduler and timer logic
in libcc to use base::TimeTicks and base::TimeDelta. This puts the units in
the type system instead of variable names and comments making second/millisecond
confusion less likely and integrates better with the rest of chromium.

BUG=


Review URL: https://chromiumcodereview.appspot.com/10956006