Ensure that the OSK does not show up at times when you click on editable fields with the mouse.
In metro mode the WM_POINTER message is sent on mouse clicks as well as opposed to desktop mode where
it only shows up when you touch the window. Fix is to set the pointer_down_context_ variable to true
in the OnTouchHandler when we receive a single touch.

BUG=150848
R=sky
Review URL: https://codereview.chromium.org/10959024