Create a new URLRequestJobFactory for isolated request contexts.
Also includes unittest from tzik@:
  http://codereview.chromium.org/10956009/

Originally the isolated URLRequestContexts used the same URLRequestJobFactory instance as the "default" request context.  This breaks filesystem: because the isolated context would incorrectly dispatch to FileSystemContext of the default URLRequestContext.

This CL makes it so the isolated contexts do not share the same URLRequestJobFactory.  There is now one URLRequestJobFactory per StoragePartition (the code equiv of one isolated context).  Note that each RequestContext and MediaRequestContext pair still share the same URLRequestJobFactory.  This is safe because they are in the isolation domain.

High level changes are:
- Each URLRequestJobFactory needs its own protocol_handler_interceptor
  which requires threading the parameter through all the
  URLRequestContext factory mess because this particular
  object must be created on the UI thread.

- GetIsolatedMediaRequestContext no longer looks up the
  app context out of the profile. Instead
  GetIsolatedMediaRequestContextGetter() does this. This
  makes it a little clearer that it is really a thin facade
  over the related isolated context.

- The common code for URLJobFactory creation is pulled
  up into SetUpJobFactoryDefaults out of both
  off_the_record and the normal profile_impl.  This will
  avoid future divergence of the setup.

- FtpProtocolHandler also moved into SetUpJobFactoryDefaults.
  Again, this is just to avoid future divergence.

- Lots of ownership passing moved to scoped_ptr<> to
  be more explicit.  No functionality change here, but lots
  of text churn.

TBR=finnur
BUG=150861

Review URL: https://codereview.chromium.org/10969017