ash: Fix app window controls on minimize/restore
We were unnecessarily tearing down and rebuilding the window frame for
apps when they are minimized.  Since the (restored) bounds don't change
we no longer do a Layout() on the transition.  We don't need to replace
the window frame in this case, so only do it when transitioning into
and out of maximized, which was the original intent of this code.

BUG=150981
TEST=manual, open file manager (Ctrl-M), click launcher icon to minimize, click launcher icon to restore, verify restore and close buttons are still visible


Review URL: https://chromiumcodereview.appspot.com/10966017