Fix toolbar keyboard accessibility on Views (alternative impl).
This broke when the location bar got moved out of the toolbar's view 
hierarchy, it no longer became possible to tab through all of the 
controls in the toolbar including the location bar. 

To fix this, I have AccessiblePaneView provide its own subclass of
FocusSearch, overriding GetParent and Contains so that the
location bar is considered to be part of the toolbar for focus
searching.

Adds a new test that should catch this type of regression in the future, 
while hopefully not being too brittle so it will still succeed as 
the toolbar changes over time. 

BUG=145835
TEST=Press F6 or Alt+Shift+T to focus the toolbar, then Tab to all of the focusable controls, make sure the location bar is included.


Review URL: https://chromiumcodereview.appspot.com/10949005