Windows: AcceleratedSurface present thread periodically polls raster status.
This is to give the threaded compositor a recent time base to work with.

Also use acknowledge with the surrent refresh period for the swap chain.

Moved call to Present to before the spin on the Query since that is unnecessary and allows earlier acknowledgement.

BUG=143854

Review URL: https://chromiumcodereview.appspot.com/10892018