aura: Start converting event-filters into event-handlers.
As a first step, add pre-target and post-target event handlers in aura::Env. The
EventFilters added in the CompoundEventFilter will be eventually converted to
EventHandlers, and installed on the Env as pre- or post-target handlers. The
RootWindow makes sure that the correct event-handlers from the env gets added to
the appropriate handlers list. This will let us get rid of the event-dispatching
logic in CompoundEventFilter.

BUG=none

Review URL: https://codereview.chromium.org/10960015