Add support for dynamically allocated port in Forwarder2.
Dynamically allocated port is necessary for running the net_unittests.


Also fixes a bug in Socket::Connect() where we forgot to wait for the notifier_fd using a select, and it was causing the forwarder to be stuck when the connection didn't complete (it did not have a timeout).

BUG=146502

Review URL: https://chromiumcodereview.appspot.com/10957043