Disable force focus call from pointer events in Chrome/Win/Desktop
This code was added in an attempt to make the on-screen-keyboard function in desktop chrome. By coincidence, it messes up popup menus/drop-down boxes. Removing the code.

Issue 105637 tracks the OSK issue. We'll need to revisit/re-enable this code when we work on that.

BUG=139420

Review URL: https://chromiumcodereview.appspot.com/10825233