Removing all explicit declarations of auto_reboot=True.
The default has been changed from False to True, so the removed lines were
redundant.

Three instances of auto_reboot=True have been left in the code. Two are used
to set the default, so they belong there. The third is in response to a bug
that remains unclosed, so I felt it wise to leave the explicit reference there.
Correct me if I'm wrong.

BUG=148767


Review URL: https://chromiumcodereview.appspot.com/10964029