Merge 158316 - Linux: Fix a wide range of crashes resulting from a very bad interaction of our dynamic menu-building code (mostly used by bookmarks, but also by some extension menus) with libdbusmenu-gtk as used in Ubuntu 12.04 for global menu support.
Basically, the problem is that it loads itself into our address space, and
shows and hides menus in ways that can't be triggered through normal user
interaction. It does this on purpose to get applications that build menus on
demand to do so; unfortunately, it doesn't work very well when the menus are
also torn down when hidden, especially the way we were doing it.

This CL changes how we tear down menus when hidden to avoid getting into loops
where menu changes are detected causing menus to be shown again, and is more
robust in the face of event orderings that don't normally occur.

BUG=131974,124110,128994
Review URL: https://codereview.chromium.org/10966043

TBR=mdm@chromium.org