Maybe fix InstantTest.NewWindowDismissesInstant on Mac.
This test times out sporadically on mac_rel (see review thread for links
to test logs). Local testing seems to suggest that this is because the
omnibox in the old browser window never loses focus, and that adding a
RunAllPending() call may fix it.

I've not been able to reproduce the original failure locally, so this is
a prospective fix, which will be confirmed by this test not timing out
on the bots anymore.

BUG=none
R=jennb@chromium.org
TEST=Watch the bots, as described above.

Review URL: https://chromiumcodereview.appspot.com/10981003