Avoid sync ipc calls for GetRootWindowRect/GetWindowRect on Android.
The sync ipc calls are expensive on Android expecially when the message loop
is busy handling both native and Java tasks. Because our simple implementation
of GetRootWindowRect/GetWindowRect in the host (both directly return
GetViewBounds() which is the same value of Rect(RenderWidget::_size)),
we are able to create short circuits of them to avoid the sync ipc calls.


Review URL: https://chromiumcodereview.appspot.com/10986030