Fix bug 152277: [Panels] Panels are still showing taskbar icons in smaller resolution on Windows
The test is added at http://codereview.chromium.org/10987037/ that contains all the test cases for more than this bug. We want to separate out the fix and the remaining complicated test such that we could merge the fix to the branch if needed.

BUG=152277
TEST=Test will be added in next patch

Review URL: https://chromiumcodereview.appspot.com/10979023