NTP5: Fixing regression bugs.
APP_LAUNCH is required by MostVisted and RecentlyClosed pages and it was originally defined in apps_page.js, which is not being used yet in NTP5. This CL moves APP_LAUNCH to new_tab.js, which seems to be the correct place for it to be once its contents are not exclusively realted to Apps the page.


BUG=

Review URL: https://chromiumcodereview.appspot.com/10969040