sync: Refactor update application
This is a refactoring patch.  It removes an unnecessary ConflictResolver
parameter from a few helper functions and rewrites the update
application logic.

This patch is part of a series.  I expect that future patches will be
easier to implement if an applicator can be reused, and the iterator
members make it difficult to do so with the current design.

BUG=147681

Review URL: https://chromiumcodereview.appspot.com/10964057