Make app launcher on Windows pinnable.
This sets the appropriate property store values so that when pinned the created shortcut will have the right icon and launch switches. A .ico has been added to the resources for this.

This also removes the need to set the window icon in the previous way, so the previous resource and code has been removed.

BUG=151672


Review URL: https://chromiumcodereview.appspot.com/10996042