The media foundation H.264 decoder which we use for H/W H.264 video decoding on Windows 7 and above has a small pool of decoding surfaces on which the decoded output is written. These surfaces are accessible via the IMFsample COM interface which we use. The IMFSample instance has ownership of the surface. When this instance is destroyed the surface is available for storing the next video frame. We save away the surface interface in the list of pending output samples.
As per above, the surfaces can be overwritten with new decoded output leading to image distortion.

A simple fix for this would be to save the IMFsample interface in the pending output sample instead
of the surface. However the ProcessOutput call blocks if it has enough output to produce a frame
and there is no available surface.

Two fixes for this:-
1. The list of pending output samples now saves away the IMFSample interface instead of the surface.

2. In the Decode call we check if we have an output frame. If yes we queue the input buffer and process
   it when the frame is released to the client.

3. The Flush implementation in the DXVA decoder has also changed to handle the above behavior. We have
   a new state kFlushing which the decoder enters. This state is checked when we process output samples.

The change in the gpu_video_decoder is to remove the OS_WIN check in the
GpuVideoDecoder::CanMoreDecodeWorkBeDone function.

The correct fix would be to not lie about being done with the input buffer until we have a valid
output frame. However this is a limitation with the media foundation decoder API.

BUG=150925
R=fischman@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10988013