Cache the channel ID
To prevent hitting rate limits on how fast we can access the server to get an
auth token to get a channel Id, we cache the channel Id when we get one.
Note that a channel Id is per profile, so we store them by profile login name.


BUG=149733


Review URL: https://chromiumcodereview.appspot.com/10951022