sync: Remove ChangeProcessor::Stop()
Stopping the ChangeProcessor from the UI is racy, especially for data
types with controllers that directly extend
NonFrontendDataTypeController.  This can lead to crashes or DCHECK
failures.

It turns out that the easiest way to fix this is to not Stop() the
ChangeProcessors from the UI thread.  There are no data types that
require this functionality.  It is at best unnecessary and at worst
harmful.

Without Stop(), there is no longer any need to check if a
ChangeProcessor is running.  Therefore, running_ and its associated
methods, DCHECKs, etc. have been removed, too.

BUG=151606


Review URL: https://chromiumcodereview.appspot.com/10979067