Temporary fix bug in disabling sync for default apps
We added a new creation_flag for default apps was_installed_by_default,
but every time we add a new creation_flag, we need to update the
installedLoader as it does not directly pick the creation_flags but
re-creates them through other preferences.

This is temporary fix for M-23 branch. A more fundamental fix is 10977048

BUG=152582


Review URL: https://chromiumcodereview.appspot.com/11034009