Don't render CSV or TSV files inline.
Add comma and tab separated values to the list of text MIME types that
will not be rendered inline. These are hard to read as a straight
textual rendering and the user almost certainly intends to open them in
a spreadsheet program or import them into some other app.

BUG=152911


Review URL: https://chromiumcodereview.appspot.com/11022006