Merge 159623 - [Chromoting] Create per-session job objects to control processes launched in a session.
Processes from different sessions cannot be attached to the same job object even in the case when their lifetimes do not overlap. The CL causes |WtsSessionProcessLauncher| to create a new job object every time it is switched to a new session. This CL also makes sure that the host process will be completely shutdown before trying to re-launch it is a different session to avoid the race with the session attach notification.

BUG=149098,153005


Review URL: https://chromiumcodereview.appspot.com/11029004

TBR=alexeypa@chromium.org
Review URL: https://codereview.chromium.org/11048007