Toggling windows with (shift)-f5 and (shift)-alt-tab does not repeat on key repeat.
While having more than one windows on the screen, when we press and hold f5 button, then the screen will flicker. In this patch, switching windows is not repeated while holding the accelerator key.

This behavior is the same as for f4 key for toggling window maximizing (recently patched).

BUG=152145
TEST=ash_unittests


Review URL: https://chromiumcodereview.appspot.com/10977074