[Android] Set a default locale for WebView.
Recent WebKit rolls have introduced an assertion for the default language that is causing
some failures in WebView. The reason is that the default locale is not properly iset in
single process mode. However, since this will probably require more elaborate solution
we'll be setting a value manually for now to unblock tests and prevent further issues.

BUG=153758
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11054012