Fix leak of MockDecryptor in ProxyDecryptorTest.
If GenerateKeyRequest() is called |real_decryptor_| is passed into |proxy_decryptor| and |proxy_decryptor| will take the ownership. But if GenerateKeyRequest() is not called the pointer is leaked. Added a scoped pointer to take the ownership before GenerateKeyRequest() is called.

Note that we still need to keep the raw pointer in the test to be able to EXPECT_CALL on the |real_decryptor_|.

TBR=scherkus@chromium.org
BUG=153786
TEST=tested locally on Linux and no memory leak.

Review URL: https://chromiumcodereview.appspot.com/11052014