Flapper field trial to use workerpool for sync file operations.
Flash private file operations are synchronous, and were being
scheduled on the browser's FILE thread.  This thread can have lots of
relatively long-running stuff on it.  Add a 50% field trial to
schedule on a blocking worker-pool thread, and histogram hung-plugin
instances to see if it helps.

BUG=153383


Review URL: https://chromiumcodereview.appspot.com/11016016