Allow extensions which can run without NPAPI to run under Windows 8 Metro.
This was previously landed as CL 10949017, but later reverted due to a failure
in ChromeOS' PFQ tests.

Extensions which include NPAPI plugins are assumed to require them, causing
them to be blocked on Windows 8 Metro mode, where NPAPI is not supported.

This CL allows extensions which include NPAPI plugins, but which can still run
without them, to load in Windows 8 Metro mode.

It also fixes the handling of extensions with no explicit NPAPI requirement.

BUG=130405


Review URL: https://chromiumcodereview.appspot.com/11050023