add wrapper to make GYP_PARALLEL=1 work for build\gyp_chromium
This is a workaround for multiprocessing on Windows. Importing in Python on
Windows doesn't search for imports that don't end in .py (and aren't
directories with an __init__.py). So, add this wrapper to avoid having
people change their command line to add a .py when running gyp_chromium.

R=dmazzoni@chromium.org

Review URL: https://codereview.chromium.org/11066021