Merge 159011 - Fix crash when trying to use a camera that is already in use.
This cl change the behaviour of VideoCaptureManager to not report errors that occur after a stream have been created. 
Ie - MediaStreamManager should not be told that starting a video capture device have failed since the video capture was not started through the MediaStreamManager. Instead - only the client that tried to start the video capture is notified of the error.

BUG=152158
TEST= See the defect report.


Review URL: https://chromiumcodereview.appspot.com/10990027

TBR=perkj@chromium.org
Review URL: https://codereview.chromium.org/11031058