Actually system GPU memory values to compute memory budget.
Previous behavior was to artifically constrain everything to the minimum spec, which resulted in corruption in high-DPI devices.

We don't have a good concept of handling multiple GPUs in the system, so I just set the budget to be the minimum across all contexts which report a value.  In principle, we could have different memory managers for each GPU (and then we'd have to handle context migration, etc, so the solution becomes very complicated).

As a reference, on a retina MBP, there are 2 GPUs -- a discrete GPU with 1GB and an integrated GPU which reports ~1.3GB, so the maximum per-tab memory budget will go from 128MB to 384MB (we still only allow a tab to have 1/2 of the total amount of memory that can be given out, to avoid excessive allocation).

BUG=153918


Review URL: https://chromiumcodereview.appspot.com/11039030