Make apks build from a single ant .xml template
The ant build files for each of the apks that we create consisted of
only fairly straightforward boilerplate. With the recent changes to
java_apk.gypi and the gyp targets we can now build these correctly from
a single .xml with the differences in how the targets are built being
reflected in the gyp files (and then passed into ant).

TBR=joth
BUG=145991


Review URL: https://chromiumcodereview.appspot.com/10959034