Merge 159954 - cros: Fix missing windows/launcher after resume/screen unlock
When the screen is locked the desktop wallpaper background widget is pulled 
"forward" to hide the user's browser windows and the launcher. There is a 
race condition between a window animation triggered by wallpaper reloading at 
device resume and code in screen unlock that "pushes" the background back 
behind the launcher and browsers. There can temporarily be two backgrounds 
and we need to move them both to the back. The extra one will be cleaned up 
when the animation completes. 

See bug for details on the race. 

I'm working on unit tests for this, but that will require some refactoring 
and I wanted to keep this patch simple for backport to R23. 

BUG=149043
TEST=interactive_ui_tests ScreenLocker.*, set option to require password on resume, sleep device (close lid or wait for timeout), resume device, verify you can unlock screen with password

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=159954

TBR=jamescook@chromium.org
Review URL: https://codereview.chromium.org/11040064