Remove dependency on cwd in license tool.
Allow tools/licenses.py to work even if the current working directory is
not the source repository root (though the current working directory is
still the default when it's run as a standalone tool). Update the
android_webview license checker to pass the repository root instead of
requiring that the cwd be the right place.

R=mnaganov@chromium.org,thakis@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11031069